+++ baybe/surrogate.py
@@ -5,8 +5,10 @@ Surrogate models, such as Gaussian processes, random forests, etc.
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Dict, Optional, Tuple, Type
+from functools import wraps
+from typing import Callable, Dict, Optional, Tuple, Type
 
+import numpy as np
 import pandas as pd
 import torch
 from botorch.fit import fit_gpytorch_model
@@ -20,21 +22,223 @@ from gpytorch.likelihoods import GaussianLikelihood
 from gpytorch.means import ConstantMean
 from gpytorch.mlls import ExactMarginalLogLikelihood
 from gpytorch.priors.torch_priors import GammaPrior
+from ngboost import NGBRegressor
+from sklearn.ensemble import RandomForestRegressor
+from sklearn.linear_model import ARDRegression
 from torch import Tensor
 
+from .scaler import DefaultScaler
 from .utils import isabstract, to_tensor
+
+MIN_TARGET_STD = 1e-6
+MIN_VARIANCE = 1e-6
+
+
+def _validate_inputs(x: Tensor) -> None:
+    """Helper function to validate the model input."""
+    if len(x) == 0:
+        raise ValueError("The model input must be non-empty.")
+
+
+def _validate_targets(y: Tensor) -> None:
+    """Helper function to validate the model targets."""
+    if y.shape[1] != 1:
+        raise NotImplementedError("The model currently supports only one target.")
+
+
+def _var_to_covar(var: Tensor) -> Tensor:
+    """
+    Converts a tensor (with optional batch dimensions) that contains (marginal)
+    variances along its last dimension into one that contains corresponding diagonal
+    covariance matrices along its last two dimensions.
+    """
+    return torch.diag_embed(var)
+
+
+def catch_constant_targets(model_cls: Type[SurrogateModel]):
+    """
+    Wraps a given `SurrogateModel` class that cannot handle constant training target
+    values such that these cases are handled by a separate model type.
+    """
+
+    class SplitModel(SurrogateModel):
+        """
+        A surrogate model that applies a separate strategy for cases where the training
+        targets are all constant and no variance can be estimated.
+        """
+
+        # Overwrite the registered subclass with the wrapped version
+        type = model_cls.type
+
+        # The posterior mode is chosen to match that of the wrapped model class
+        joint_posterior = model_cls.joint_posterior
+
+        def __init__(self, *args, **kwargs):
+            """Stores an instance of the underlying model class."""
+            self.model = model_cls(*args, **kwargs)
+
+        def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+            """Calls the posterior function of the internal model instance."""
+            mean, var = self.model._posterior(  # pylint: disable=protected-access
+                candidates
+            )
+
+            # If a joint posterior is expected but the model has been overriden by one
+            # that does not provide covariance information, construct a diagonal
+            # covariance matrix
+            if self.joint_posterior and not self.model.joint_posterior:
+                var = _var_to_covar(var)
+
+            return mean, var
+
+        def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
+            """Selects a model based on the variance of the targets and fits it."""
+
+            # https://github.com/pytorch/pytorch/issues/29372
+            # Needs 'unbiased=False' (otherwise, the result will be NaN for scalars)
+            if torch.std(train_y.ravel(), unbiased=False) < MIN_TARGET_STD:
+                self.model = MeanPredictionModel(self.model.searchspace)
+
+            # Fit the selected model with the training data
+            self.model.fit(train_x, train_y)
+
+        def __getattribute__(self, attr):
+            """
+            Accesses the attributes of the class instance if available, otherwise uses
+            the attributes of the internal model instance.
+            """
+            # Try to retrieve the attribute in the class
+            try:
+                val = super().__getattribute__(attr)
+            except AttributeError:
+                pass
+            else:
+                return val
+
+            # If the attribute has not been overwritten, use that of the internal model
+            return self.model.__getattribute__(attr)
+
+    return SplitModel
+
+
+def scale_model(model_cls: Type[SurrogateModel]):
+    """
+    Wraps a given `SurrogateModel` class such that it operates with scaled
+    representations of the training and test data.
+    """
+
+    class ScaledModel(model_cls):
+        """Overrides the methods of the given model class such the use scaled data."""
+
+        def __init__(self, *args, **kwargs):
+            """Stores an instance of the underlying model class and a scaler object."""
+            self.model = model_cls(*args, **kwargs)
+            self.scaler = None
+
+        def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+            """
+            Calls the posterior function of the internal model instance on
+            a scaled version of the test data and rescales the output accordingly.
+            """
+            candidates = self.scaler.transform(candidates)
+            mean, covar = self.model._posterior(  # pylint: disable=protected-access
+                candidates
+            )
+            return self.scaler.untransform(mean, covar)
+
+        def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
+            """Fits the scaler and the model using the scaled training data."""
+            self.scaler = DefaultScaler(self.model.searchspace)
+            train_x, train_y = self.scaler.fit_transform(train_x, train_y)
+            self.model.fit(train_x, train_y)
+
+        def __getattribute__(self, attr):
+            """
+            Accesses the attributes of the class instance if available, otherwise uses
+            the attributes of the internal model instance.
+            """
+            # Try to retrieve the attribute in the class
+            try:
+                val = super().__getattribute__(attr)
+            except AttributeError:
+                pass
+            else:
+                return val
+
+            # If the attribute has not been overwritten, use that of the internal model
+            return self.model.__getattribute__(attr)
+
+    return ScaledModel
+
+
+def batchify(
+    posterior: Callable[[SurrogateModel, Tensor], Tuple[Tensor, Tensor]]
+) -> Callable[[SurrogateModel, Tensor], Tuple[Tensor, Tensor]]:
+    """
+    Wraps `SurrogateModel` posterior functions that are incompatible with t- and
+    q-batching such that they become able to process batched inputs.
+    """
+
+    @wraps(posterior)
+    def sequential_posterior(
+        model: SurrogateModel, candidates: Tensor
+    ) -> [Tensor, Tensor]:
+        """A posterior function replacement that processes batches sequentially."""
+
+        # If no batch dimensions are given, call the model directly
+        if candidates.ndim == 2:
+            return posterior(model, candidates)
+
+        # Keep track of batch dimensions
+        t_shape = candidates.shape[:-2]
+        q_shape = candidates.shape[-2]
+
+        # If the posterior function provides full covariance information, call it
+        # t-batch by t-batch
+        if model.joint_posterior:  # pylint: disable=no-else-return
+
+            # Flatten all t-batch dimensions into a single one
+            flattened = candidates.flatten(end_dim=-3)
+
+            # Call the model on each (flattened) t-batch
+            out = (posterior(model, batch) for batch in flattened)
+
+            # Collect the results and restore the batch dimensions
+            mean, covar = zip(*out)
+            mean = torch.reshape(torch.stack(mean), t_shape + (q_shape,))
+            covar = torch.reshape(torch.stack(covar), t_shape + (q_shape, q_shape))
+
+            return mean, covar
+
+        # Otherwise, flatten all t- and q-batches into a single q-batch dimension
+        # and evaluate the posterior function in one go
+        else:
+
+            # Flatten *all* batches into the q-batch dimension
+            flattened = candidates.flatten(end_dim=-2)
+
+            # Call the model on the entire input
+            mean, var = posterior(model, flattened)
+
+            # Restore the batch dimensions
+            mean = torch.reshape(mean, t_shape + (q_shape,))
+            var = torch.reshape(var, t_shape + (q_shape,))
+
+            return mean, var
+
+    return sequential_posterior
 
 
 class SurrogateModel(ABC):
     """Abstract base class for all surrogate models."""
 
-    # TODO: to support other models than GPs, an interface to botorch's acquisition
-    #  functions must be created (e.g. via a dedicated 'predict' method)
-
-    type: str
+    # Dictionary for bookkeeping of subclasses
     SUBCLASSES: Dict[str, Type[SurrogateModel]] = {}
 
-    @abstractmethod
+    # Class properties
+    type: str
+    joint_posterior: bool
+
     def posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
         """
         Evaluates the surrogate model at the given candidate points.
@@ -53,10 +257,49 @@ class SurrogateModel(ABC):
             The posterior means and posterior covariance matrices of the t-batched
             candidate points.
         """
+        # Evaluate the posterior distribution
+        mean, covar = self._posterior(candidates)
+
+        # Apply covariance transformation for marginal posterior models
+        if not self.joint_posterior:
+            covar = _var_to_covar(covar)
+
+        # Add small diagonal variances for numerical stability
+        covar.add_(torch.eye(covar.shape[-1]) * MIN_VARIANCE)
+
+        return mean, covar
 
     @abstractmethod
+    def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+        """
+        Implements the actual posterior evaluation logic. In contrast to its public
+        counterpart, no data validation/transformation is carried out but only the raw
+        posterior computation is conducted.
+
+        Note:
+        -----
+        The public `posterior` method *always* returns a full covariance matrix. By
+        contrast, this method may return either a covariance matrix or a tensor
+        of marginal variances, depending on the models `joint_posterior` flag. The
+        optional conversion to a covariance matrix is handled by the public method.
+        """
+
     def fit(self, train_x: Tensor, train_y: Tensor) -> None:
         """Trains the surrogate model on the provided data."""
+
+        # Validate the training data
+        _validate_inputs(train_x)
+        _validate_targets(train_y)
+
+        return self._fit(train_x, train_y)
+
+    @abstractmethod
+    def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
+        """
+        Implements the actual fitting logic. In contrast to its public counterpart,
+        no data validation/transformation is carried out but only the raw fitting
+        operation is conducted.
+        """
 
     @classmethod
     def __init_subclass__(cls, **kwargs):
@@ -70,6 +313,7 @@ class GaussianProcessModel(SurrogateModel):
     """A Gaussian process surrogate model."""
 
     type = "GP"
+    joint_posterior = True
 
     def __init__(self, searchspace: pd.DataFrame):
         self.model: Optional[SingleTaskGP] = None
@@ -78,19 +322,13 @@ class GaussianProcessModel(SurrogateModel):
         #  DataFrame
         self.searchspace = searchspace
 
-    def posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+    def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
         """See base class."""
         posterior = self.model.posterior(candidates)
         return posterior.mvn.mean, posterior.mvn.covariance_matrix
 
-    def fit(self, train_x: Tensor, train_y: Tensor) -> None:
+    def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
         """See base class."""
-
-        # validate input
-        if len(train_x) == 0:
-            raise ValueError("The training data set must be non-empty.")
-        if train_y.shape[1] != 1:
-            raise NotImplementedError("The model currently supports only one target.")
 
         # get the input bounds from the search space
         searchspace = to_tensor(self.searchspace)
@@ -177,3 +415,139 @@ class GaussianProcessModel(SurrogateModel):
         )
         mll = ExactMarginalLogLikelihood(self.model.likelihood, self.model)
         fit_gpytorch_model(mll, optimizer=fit_gpytorch_torch, options={"disp": False})
+
+
+class MeanPredictionModel(SurrogateModel):
+    """
+    A trivial surrogate model that provides the average value of the training targets
+    as posterior mean and a (data-independent) constant posterior variance.
+    """
+
+    type = "MP"
+    joint_posterior = False
+
+    def __init__(self, searchspace: pd.DataFrame):  # pylint: disable=unused-argument
+        self.target_value = None
+
+    @batchify
+    def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+        """See base class."""
+        # TODO: use target value bounds for covariance scaling when explicitly provided
+        mean = self.target_value * torch.ones([len(candidates)])
+        var = torch.ones(len(candidates))
+        return mean, var
+
+    def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
+        """See base class."""
+        self.target_value = train_y.mean().item()
+
+
+@catch_constant_targets
+@scale_model
+class RandomForestModel(SurrogateModel):
+    """A random forest surrogate model."""
+
+    type = "RF"
+    joint_posterior = False
+
+    def __init__(self, searchspace: pd.DataFrame):
+        self.model: Optional[RandomForestRegressor] = None
+        # TODO: the surrogate model should work entirely on Tensors (parameter name
+        #  agnostic) -> the scaling information should not be provided in form of a
+        #  DataFrame
+        self.searchspace = searchspace
+
+    @batchify
+    def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+        """See base class."""
+
+        # Evaluate all trees
+        # NOTE: explicit conversion to ndarray is needed due to a pytorch issue:
+        # https://github.com/pytorch/pytorch/pull/51731
+        # https://github.com/pytorch/pytorch/issues/13918
+        predictions = torch.from_numpy(
+            np.asarray(
+                [
+                    self.model.estimators_[tree].predict(candidates)
+                    for tree in range(self.model.n_estimators)
+                ]
+            )
+        )
+
+        # Compute posterior mean and variance
+        mean = predictions.mean(dim=0)
+        var = predictions.var(dim=0)
+
+        return mean, var
+
+    def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
+        """See base class."""
+        self.model = RandomForestRegressor()
+        self.model.fit(train_x, train_y.ravel())
+
+
+@catch_constant_targets
+@scale_model
+class NGBoostModel(SurrogateModel):
+    """A natural-gradient-boosting surrogate model."""
+
+    type = "NG"
+    joint_posterior = False
+
+    def __init__(self, searchspace: pd.DataFrame):
+        self.model: Optional[NGBRegressor] = None
+        # TODO: the surrogate model should work entirely on Tensors (parameter name
+        #  agnostic) -> the scaling information should not be provided in form of a
+        #  DataFrame
+        self.searchspace = searchspace
+
+    @batchify
+    def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+        """See base class."""
+        # Get predictions
+        dists = self.model.pred_dist(candidates)
+
+        # Split into posterior mean and variance
+        mean = torch.from_numpy(dists.mean())
+        var = torch.from_numpy(dists.var)
+
+        return mean, var
+
+    def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
+        """See base class."""
+        self.model = NGBRegressor(n_estimators=25, verbose=False).fit(
+            train_x, train_y.ravel()
+        )
+
+
+@catch_constant_targets
+@scale_model
+class BayesianLinearModel(SurrogateModel):
+    """A Bayesian linear regression surrogate model."""
+
+    type = "BL"
+    joint_posterior = False
+
+    def __init__(self, searchspace: pd.DataFrame):
+        self.model: Optional[ARDRegression] = None
+        # TODO: the surrogate model should work entirely on Tensors (parameter name
+        #  agnostic) -> the scaling information should not be provided in form of a
+        #  DataFrame
+        self.searchspace = searchspace
+
+    @batchify
+    def _posterior(self, candidates: Tensor) -> Tuple[Tensor, Tensor]:
+        """See base class."""
+        # Get predictions
+        dists = self.model.predict(candidates.numpy(), return_std=True)
+
+        # Split into posterior mean and variance
+        mean = torch.from_numpy(dists[0])
+        var = torch.from_numpy(dists[1]).pow(2)
+
+        return mean, var
+
+    def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
+        """See base class."""
+        self.model = ARDRegression()
+        self.model.fit(train_x, train_y.ravel())